Fix Umami activation
