+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "schedules",
-    "version": "2.2.0-edge.9",
+    "version": "2.2.1-edge.1",
     "private": true,
     "scripts": {
         "serve": "vue-cli-service serve",
